5 past days view array created
