+++ README.md
@@ -1,6 +1,6 @@
 # Med-Tracker
 
-[Live Link](https://sjimmykang-med-tracker.netlify.app/)
+## [Live Link](https://sjimmykang-med-tracker.netlify.app/)
 ## [DEV Live Link](https://sjimmykang-med-tracker-development.netlify.app/)
 
 # Getting Started with Create React App
