+++ public/index.html
@@ -2,6 +2,10 @@
 <html lang="en">
   <head>
     <meta charset="utf-8" />
+    <!-- 
+      Favicon and logo192, logo512 from: https://www.svgrepo.com/svg/389888/pill
+      by: wishforge.games
+    -->
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
